keymaps: split key mappings to setup
