+++ lua/overlays/rc/dial.lua
@@ -1,0 +1,25 @@
+local augend = require("dial.augend")
+require("dial.config").augends:register_group({
+  -- default augends used when no group name is specified
+  default = {
+    augend.integer.alias.decimal, -- nonnegative decimal number (0, 1, 2, 3, ...)
+    augend.integer.alias.hex, -- nonnegative hex number  (0x01, 0x1a1f, etc.)
+    augend.date.alias["%Y/%m/%d"], -- date (2022/02/19, etc.)
+    augend.date.alias["%Y-%m-%d"],
+    augend.date.alias["%m/%d"],
+    augend.date.alias["%H:%M"],
+    augend.constant.alias.bool, -- boolean value (true <-> false)
+    -- switch between and/or &&/||
+    augend.constant.new({
+      elements = { "and", "or" },
+      word = true,
+      cyclic = true,
+    }),
+    augend.constant.new({
+      elements = { "&&", "||" },
+      word = false,
+      cyclic = true,
+    }),
+    augend.semver.alias.semver,
+  },
+})
